[BEAM-6188] Separate source delay logic from the IO
